requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
et-xmlfile==1.0.1
idna==2.8
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.2
pycodestyle==2.5.0
pylint==2.4.4
pymongo==3.10.0
pymssql==2.1.4
PyMySQL==0.9.3
python-dateutil==2.8.1
redis==3.3.11
requests==2.22.0
selenium==3.141.0
six==1.13.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
